fix(fix-repository-owner): add testing workflow
add testing workflow
