+++ src/createComfyRegistryPullRequests.ts
@@ -1,9 +1,10 @@
 import minimist from "minimist";
 import pMap from "p-map";
-import { clone_modify_push_Branches } from "./clone_modify_push_Branches";
 import { createGithubForkForRepo } from "./createGithubForkForRepo";
 import { createGithubPullRequest } from "./createGithubPullRequest";
 import type { GithubPull } from "./gh/GithubPull";
+import { makePublishcrBranch } from "./makePublishBranch";
+import { makePyprojectBranch } from "./makeTomlBranch";
 import { parsePulls } from "./parsePullsState";
 if (import.meta.main) {
   // test repo
@@ -11,6 +12,7 @@ if (import.meta.main) {
   const test_repo = minimist(process.argv.slice(2)).argv || default_test_repo;
   console.info(await createComfyRegistryPullRequests(test_repo));
 }
+
 export async function createComfyRegistryPullRequests(upstreamRepoUrl: string) {
   console.log("forking " + upstreamRepoUrl);
   const forkedRepo = await createGithubForkForRepo(upstreamRepoUrl);
@@ -22,4 +24,27 @@ export async function createComfyRegistryPullRequests(upstreamRepoUrl: string) {
   console.log("Registry PRs DONE");
 
   return ([...prs] as GithubPull[]).map((e) => parsePulls([e])[0]);
+}
+
+
+/**
+ * this function will:
+ *
+ * for each branch (pyproject, publishcr)
+ * 1. clone forked srepo
+ * 2. modify on forked repo (add pyproject.toml, add .github/workflow/publish.yml)
+ * 3. commit changes, push to forked repo
+ * 4. check if all branch is ready, create PR to upstream repo
+ */
+export async function clone_modify_push_Branches(upstreamUrl: string, forkUrl: string) {
+  const pyprojectBranchInfo = makePyprojectBranch(upstreamUrl, forkUrl);
+  const publishcrBranchInfo = makePublishcrBranch(upstreamUrl, forkUrl);
+  return (await Promise.all([pyprojectBranchInfo, publishcrBranchInfo])).map(({ body, branch, title, type }) => ({
+    body,
+    branch,
+    title,
+    type,
+    srcUrl: forkUrl,
+    dstUrl: upstreamUrl,
+  }));
 }
